pyproject.toml
[project]
name = "agent-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.121.2",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "openai-agents>=0.1.0",
    "requests>=2.31.0",
]
